pyproject.toml
[tool.poetry]
name = "platzkey"
version = "0.1.0"
description = "Another blog in python"
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
humanize = "^4.3.0"
Flask = "^2.2.2"
PyYAML = "^6.0"
Flask-Babel = "^2.0.0"
Flask-WTF = "^1.0.1"
gql = "^3.4.0"
aiohttp = "^3.8.3"
Flask-Markdown = "^0.3"

[tool.poetry.group.dev.dependencies]
pytest = "^7.1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
